refactor: move component to common components directory
