+++ browser/components/SnippetTab.js
