+++ browser/components/SnippetTab.styl
